refactoring: splitting into module
